+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Integrated the loss data files to be part of deus, so
   the user does not have to give to the calls for deus anymore
+- Added aliases for intensity measurements
 
 # 2019-09-18